Adapt to our new API guidelines
This commit refactors the Rails structure to comply to our new API
development guidelines.

We adopt a unique URL scheme for HTML and JSON requests but we process
them in separate Rails controllers to avoid having to adapt the
behaviour to the format in every single controller action.

We also re-organize the views file structure to match the new
controllers organization while keeping compliance with Rails naming
scheme used by Rails autload.